--- libs/options-api/src/main/java/gov/cdc/nbs/option/name/type/NameTypeOptionFinder.java
@@ -1,25 +1,0 @@
-package gov.cdc.nbs.option.name.type;
-
-import gov.cdc.nbs.option.jdbc.SQLBasedOptionFinder;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.stereotype.Component;
-
-@Component
-class NameTypeOptionFinder extends SQLBasedOptionFinder {
-
-  // private static final String QUERY = """
-  //     select
-  //         code                                as [value],
-  //         code_short_desc_txt                 as [name],
-  //         row_number() over (order by [code]) as [order]
-  //     from NBS_SRTE..Code_value_general
-  //     where code_set_nm = :set_name
-  //     order by
-  //       code
-  //     """;
-
-  NameTypeOptionFinder(final JdbcTemplate template) {
-    super(QUERY, template);
-  }
-
-}
